La ficha del préstamo cabía en dos ficheros y ya no
`DetallePrestamo.module.css` estaba en 799 líneas y `DetallePrestamoPage.tsx`
en 793, las dos a un renglón del techo del trinquete, y con la pantalla en
plena remodelación: el siguiente cambio, por pequeño que fuera, habría
tenido que empezar por recortar comentarios para hacerle sitio.

Dos cortes, ninguno arbitrario:

· Las clases `bonif*` las usa un solo componente, así que se van con él a
  `TarjetaBonificaciones.module.css`. Lo COMPARTIDO —`card`, `cardT`,
  `mono`, `teorico`— se queda donde estaba y el componente importa los dos:
  duplicarlo sería tener dos definiciones del mismo borde esperando a
  separarse.

· El hero sale a `HeroPrestamo.tsx`. Es la mitad de la pantalla que NO
  cambia con el tipo de préstamo —siempre son las mismas seis cifras—
  frente a la rejilla de abajo, que sí se adapta.

Queda 500 / 215 / 295 líneas de TSX y 700 / 112 de CSS.

Sin cambio de comportamiento, y verificado como tal: el JSX extraído es
idéntico al que había salvo las cuatro sustituciones de props
(`setAmortizando`→`onAmortizar`, `navigate`→`onEditar`, `hoy`→`anio`,
`fiscalidad.pctDeducible`→`pctDeducible`), comprobado con un diff contra el
original. Y todas las clases referenciadas en la carpeta resuelven contra
el fichero del que se importan — que es la comprobación que hace falta al
partir un CSS module, porque una referencia rota compila en silencio.
