+++ src/modules/financiacion/vista/DetallePrestamoPage.tsx
@@ -36,7 +36,7 @@ import {
   getProximaRevision,
   getTinVigente,
 } from '../../../services/prestamos/lecturas';
-import { cuadroSeguroDe, metaDestino } from './datos';
+import { cuadroSeguroDe } from './datos';
 import { useRevisionPendiente } from './useRevisionPendiente';
 import { simulacionRevision } from '../../../services/prestamos/simulacionRevision';
 import { calendarioDe, proximaRevision } from '../../../services/bonificaciones/revisionDelBanco';
@@ -45,6 +45,7 @@ import { inmueblesAlquiladosEn } from '../../../services/prestamos/inmueblesAlqu
 import { getFinancialValuesSnapshot } from '../../../services/financialValuesService';
 import CuadroCompleto from './CuadroCompleto';
 import CuadroPreview from './CuadroPreview';
+import HeroPrestamo from './HeroPrestamo';
 import TarjetaBonificaciones from './TarjetaBonificaciones';
 import type { Prestamo } from '../../../types/prestamos';
 import {
@@ -59,8 +60,6 @@ import {
   enmascarado,
   ETIQUETA_TIPO,
   diaMesAnio,
-  eurAFavor,
-  eurDeuda,
   eurPlano,
   mesAnio,
   pct,
@@ -76,16 +75,6 @@ const hoyISO = (): string => {
   return `${d.getFullYear()}-${mm}-${dd}`;
 };
 
-const Segmentos: React.FC<{ pct: number }> = ({ pct: valor }) => {
-  const encendidos = Math.round(Math.min(100, Math.max(0, valor)) / 10);
-  return (
-    <div className={styles.heroAmort} aria-hidden>
-      {Array.from({ length: 10 }, (_, i) => (
-        <i key={i} className={i < encendidos ? styles.pipOn : undefined} />
-      ))}
-    </div>
-  );
-};
 
 const DetallePrestamoPage: React.FC = () => {
   const navigate = useNavigate();
@@ -290,139 +279,18 @@ const DetallePrestamoPage: React.FC = () => {
         <span className={styles.migaActual}>{prestamo.nombre || 'Préstamo'}</span>
       </div>
 
-      {/* ── HERO navy · el nombre DENTRO, como la vista principal ── */}
-      <div className={styles.hero}>
-      <div className={styles.heroHead}>
-        <div className={styles.heroHeadIzq}>
-          <h1 className={styles.heroTitulo}>{prestamo.nombre || 'Préstamo sin nombre'}</h1>
-          <div className={styles.heroSub}>
-            {ETIQUETA_TIPO[prestamo.tipo] ?? 'préstamo'} · destino{' '}
-            <span className={styles.mono}>{metaDestino(prestamo)}</span>
-            {numContrato && (
-              <>
-                {' · nº '}
-                <span className={styles.mono}>{numContrato}</span>
-              </>
-            )}
-          </div>
-        </div>
-        <div className={styles.heroAcciones}>
-          {/* No hay pantalla ni documento de FEIN enlazable todavía · el botón
-              se deja visible y deshabilitado en vez de inventar un destino. */}
-          <button type="button" className={styles.btnGhostNavy} disabled title="Pendiente · no hay FEIN enlazada">
-            <Icons.Contratos size={14} strokeWidth={2} />
-            Ver FEIN
-          </button>
-          {/* Tercera acción · no está en el mockup, pero adelantar capital es
-              una operación real que ya funciona y el Detalle era su única
-              puerta. Decisión de Jose · se mantiene aquí hasta que Mi Plan /
-              Acelerar le dé sitio propio. */}
-          <button type="button" className={styles.btnGhostNavy} onClick={() => setAmortizando(true)}>
-            <Icons.Amortizar size={14} strokeWidth={2} />
-            Amortizar
-          </button>
-          <button
-            type="button"
-            className={styles.btnOroNavy}
-            onClick={() => navigate(`/financiacion/${prestamo.id}/editar`)}
-          >
-            <Icons.Edit size={14} strokeWidth={2} />
-            Editar
-          </button>
-        </div>
-      </div>
-        <div className={styles.hkRow}>
-        <div className={styles.hk}>
-          <div className={styles.hkLab}>Capital vivo</div>
-          <div className={styles.hkVal}>{eurDeuda(datos.capitalVivo)}</div>
-          <div className={styles.hkSub}>
-            de <span className={styles.mono}>{eurPlano(datos.principalInicial)}</span> inicial
-          </div>
-        </div>
-
-        <div className={styles.hk}>
-          <div className={styles.hkLab}>Cuota{esMixtoOVariable ? ' actual' : ''}</div>
-          <div className={styles.hkVal}>{eurDeuda(datos.cuota)}</div>
-          <div className={styles.hkSub}>
-            {datos.revision ? (
-              <span className={styles.hkCambio}>
-                {datos.revision.cuotaDespues > datos.revision.cuotaAntes ? '↑' : '↓'}{' '}
-                {eurPlano(datos.revision.cuotaDespues)} desde{' '}
-                {mesAnio(datos.revision.fecha).split(' ')[0]}
-              </span>
-            ) : (
-              'constante · sin revisiones'
-            )}
-          </div>
-        </div>
-
-        <div className={styles.hk}>
-          {/* «TIN BONIFICADO 2,60 %» era la misma mentira que la tarjeta, en el
-              número más grande de la pantalla: en el tramo fijo de esta mixta
-              no hay nada bonificado — el 2,600 % es el tipo a secas. Lo decide
-              `rebajanHoy`, la misma respuesta que usa el cuadro. */}
-          <div className={styles.hkLab}>
-            TIN{' '}
-            {bonificaciones.rebajanHoy && bonificaciones.rebajaTotal > 0
-              ? 'bonificado'
-              : prestamo.tipo === 'FIJO'
-                ? 'fijo'
-                : 'vigente'}
-          </div>
-          <div className={styles.hkVal}>{pct(datos.tin)}</div>
-          <div className={styles.hkSub}>
-            {datos.revision &&
-            Math.abs(datos.revision.tinDespues - datos.revision.tinAntes) >= 0.005 ? (
-              <>
-                → <span className={styles.mono}>{pct(datos.revision.tinDespues).replace(' %', '')}</span>{' '}
-                tras revisión
-              </>
-            ) : prestamo.tipo === 'FIJO' ? (
-              'toda la vida del préstamo'
-            ) : (
-              'sin revisión a la vista'
-            )}
-          </div>
-        </div>
-
-        <div className={styles.hk}>
-          <div className={styles.hkLab}>Amortizado</div>
-          <Segmentos pct={datos.pctAmortizado} />
-          <div className={styles.hkSub}>
-            <span className={styles.mono}>{Math.round(datos.pctAmortizado)}%</span> ·{' '}
-            <span className={styles.mono}>{datos.progreso.pagadas}</span> de{' '}
-            <span className={styles.mono}>{datos.progreso.total}</span> cuotas
-          </div>
-        </div>
-
-        <div className={styles.hk}>
-          <div className={styles.hkLab}>Vence</div>
-          <div className={`${styles.hkVal} ${styles.oro}`}>{mesAnio(datos.vencimiento)}</div>
-          <div className={styles.hkSub}>
-            quedan <span className={styles.mono}>{datos.progreso.restantes}</span> cuotas
-          </div>
-        </div>
-
-        <div className={styles.hk}>
-          <div className={styles.hkLab}>
-            {mostrarDeducible ? `Deducible ${hoy.slice(0, 4)}` : 'Interés pendiente'}
-          </div>
-          <div className={styles.hkVal}>
-            {mostrarDeducible ? eurAFavor(datos.deducible) : eurDeuda(datos.interesPendiente)}
-          </div>
-          <div className={styles.hkSub}>
-            {mostrarDeducible ? (
-              <>
-                <span className={styles.mono}>{Math.round(fiscalidad.pctDeducible)}%</span> · casilla
-                0105
-              </>
-            ) : (
-              'lo que queda de coste'
-            )}
-          </div>
-        </div>
-        </div>
-      </div>
+      <HeroPrestamo
+        prestamo={prestamo}
+        datos={datos}
+        bonificaciones={bonificaciones}
+        anio={hoy.slice(0, 4)}
+        numContrato={numContrato}
+        mostrarDeducible={mostrarDeducible}
+        pctDeducible={fiscalidad.pctDeducible}
+        esMixtoOVariable={esMixtoOVariable}
+        onAmortizar={() => setAmortizando(true)}
+        onEditar={() => navigate(`/financiacion/${prestamo.id}/editar`)}
+      />
 
       {/* ── Rejilla 2×2 ── */}
       <div className={styles.rejilla}>
