+++ src/modules/financiacion/vista/TarjetaBonificaciones.module.css
@@ -1,0 +1,112 @@
+/* ═══════════════════════════════════════════════════════════════
+   Bonificaciones · la tarjeta de lo que ganas para el periodo que viene
+   ═══════════════════════════════════════════════════════════════
+
+   Salieron de `DetallePrestamo.module.css` cuando ese fichero llegó a 799
+   líneas, a un renglón del techo del trinquete, y con la ficha en plena
+   remodelación. El corte no es arbitrario: estas clases las usa un solo
+   componente, `TarjetaBonificaciones`.
+
+   Lo COMPARTIDO —`card`, `cardT`, `cardNota`, `mono`, `teorico`, `flecha`—
+   se queda donde estaba y el componente importa los dos ficheros: son el
+   armazón de todas las tarjetas de la ficha, y duplicarlo aquí sería tener
+   dos definiciones del mismo borde esperando a separarse.
+   ═══════════════════════════════════════════════════════════════ */
+
+.bonifLista {
+  min-height: 0;
+  overflow-y: auto;
+}
+
+.bonif {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: var(--atlas-v5-sp-2);
+  padding: 7px 0;
+  border-bottom: 1px solid var(--atlas-v5-line-3);
+  font-size: 12.5px;
+}
+
+.bonif:last-of-type { border-bottom: none; }
+
+.bonifIzq {
+  display: flex;
+  align-items: center;
+  gap: 9px;
+  color: var(--atlas-v5-ink-2);
+  min-width: 0;
+}
+
+.bonifCheck,
+.bonifPendiente {
+  width: 18px;
+  height: 18px;
+  border-radius: 5px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-shrink: 0;
+}
+
+.bonifCheck {
+  background: var(--atlas-v5-gold-wash);
+  color: var(--atlas-v5-gold-ink);
+}
+
+/* La que aún no tienes se ve, apagada · esconderla diría que no hay nada
+   que ir a buscar, y es justo la que puedes conseguir. */
+.bonifPendiente {
+  background: var(--atlas-v5-card-alt);
+  border: 1px solid var(--atlas-v5-line);
+}
+
+.bonifApagada { color: var(--atlas-v5-ink-4); }
+
+.bonifVal,
+.bonifValApagado {
+  font-family: var(--atlas-v5-font-mono-num);
+  font-variant-numeric: tabular-nums;
+  font-weight: var(--atlas-v5-fw-bold);
+  white-space: nowrap;
+}
+
+.bonifVal { color: var(--atlas-v5-ink); }
+.bonifValApagado { color: var(--atlas-v5-ink-5); }
+
+.bonifPie {
+  margin-top: auto;
+  padding-top: 11px;
+  border-top: 1px solid var(--atlas-v5-line-2);
+  flex-shrink: 0;
+}
+
+.bonifPieFila {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  font-size: var(--atlas-v5-fs-sub-sm);
+  color: var(--atlas-v5-ink-3);
+}
+
+.bonifEfecto {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: var(--atlas-v5-sp-2);
+  margin-top: 9px;
+  padding: var(--atlas-v5-sp-2) 11px;
+  background: var(--atlas-v5-gold-wash);
+  border-radius: var(--atlas-v5-radius-icon);
+  font-size: var(--atlas-v5-fs-sub-sm);
+  color: var(--atlas-v5-gold-ink);
+  font-weight: var(--atlas-v5-fw-semibold);
+}
+
+.bonifEfecto .mono { font-weight: var(--atlas-v5-fw-bold); }
+
+.flecha {
+  font-family: var(--atlas-v5-font-mono-num);
+  font-weight: var(--atlas-v5-fw-bold);
+}
+
